All Errors handled by file validators
This changes the `generateMarkerEdges` function to treat missing
target files or markers as broken edges rather than reporting them
directly. The validators have then been updated to understand this
kind of broken edge and report them. This will allow for other kinds
of validators to handle this kind of error with a more helpful message
(such as those that could be used to support linters).

This also addresses an issue that Flow can't catch due to poor core
typing. It turns out that regex optional groups can come back as
undefined in a match, even though the type is a string. So, now
the two optional groups in our tag parsing are coerced to empty strings
in the case that they aren't present (which is what was expected).