Major reworking to separate parsing from processing
This makes it so that the ErrorDetail is the result of parsing and it is that which the processing occurs on. The processing itself is now done by the OutputSink, which handles all the state of processing errors appropriately and outputting the right things.

Only thing not done yet is the JSON output, which will be different than before and just reuse the ErrorDetails format rather than another JSON format. If we find we need to output more fields, we can likely update the error details object to include that data (just as the MarkerEdge once did).

TODO: Fix tests and add JSON output