[somewhatabstract/issue2039] Add cache support (#2040)
## Summary:
This adds some new arguments so that we can split out the parsing of tags (aka markers) from files, and the processing of the markers. This will speed up our integration tests, but potentially opens the door for other uses. For example, the upcoming migration tool could be run using a cache so that we can re-run it with the same cache without having to re-parse all the same files, which will make it easier for me to test against webapp's complex tagging without having to wait to parse all the files every time.

This is not some cache to speed up regular checksync usage (like, say, how babel uses a cache). That would take a lot more effort (like working out cache invalidation, etc.) and isn't planned at this time.

The output cache does not include the settings used to create it, so it should be run against the same config or CLI args as it was created.

Issue: FEI-6294 and closes #2039

## Test plan:
I have updated the integration tests to use this new mode so that we only do the parsing once per example, regardless of the scenarios. This should help with testing the migration tool that I'm working on to help move from the local sync tags to the remote sync tags where needed.

`pnpm test`

Author: somewhatabstract

Reviewers: somewhatabstract, kevinb-khan, jeresig, jeremywiebe, kevinbarabash

Required Reviewers:

Approved By: kevinb-khan

Checks: ✅ 11 checks were successful, ⏭️  1 check has been skipped

Pull Request URL: https://github.com/somewhatabstract/checksync/pull/2040