[fei6294.2.errorcodes] Move error creation to helper functions (#2050)
## Summary:
This is a tidy-up task before I add the last bit of migration stuff. It isn't perfect but it moves the error creation to helper functions to make things a little more consistent and help spot patterns.

Trying to work out the smallest set of args to send was a little trying so I didn't try too hard on this pass. I expect a bigger refactor will be needed to make this nicer; such as not including full text messages in the details but letting those be calculated at a later stage. Some clear areas of confusion are around when a target path is used for the tag itself (needs to be relative to the root marker), or when it is used to report to the user (needs to be relative to the current working directory).

Issue: FEI-6294

## Test plan:
`pnpm test` to show that things are good.

Author: somewhatabstract

Reviewers: jaredly, somewhatabstract, jeresig, kevinbarabash

Required Reviewers:

Approved By: jaredly

Checks: ✅ 11 checks were successful, ⏭️  1 check has been skipped

Pull Request URL: https://github.com/somewhatabstract/checksync/pull/2050