[🔥AUDIT🔥] Remove arg default for includeDotPaths (#2120)
🖍 _This is an audit!_ 🖍

## Summary:
This was preventing `checksync --help` from working, as the default is seen as the arg being used and so then it says it can't be used with `--help`. We want to still prevent them being combined as it doesn't make sense, so we remove the default for the arg and instead rely on our internal default logic to handle setting that default value. This is what we want anyway as we don't want the config to be overridden by the default arg.

Issue: XXX-XXXX

## Test plan:
`pnpm test`

Author: somewhatabstract

Auditors:

Required Reviewers:

Approved By:

Checks: ⏭️  1 check has been skipped, ✅ 1 check was successful, ⌛ 2 checks are pending

Pull Request URL: https://github.com/somewhatabstract/checksync/pull/2120